requirements.txt
amqp==2.6.1
asgiref==3.3.4
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.2
Deprecated==1.2.13
Django==3.1.7
django-braces==1.14.0
django-debug-toolbar==3.2
django-extensions==3.1.2
django-redis==5.0.0
djangorestframework==3.12.4
drf-writable-nested==0.6.3
importlib-metadata==4.11.3
install==1.3.4
jsonfield==3.1.0
kombu==4.6.11
model-bakery==1.3.1
packaging==21.3
Pillow==7.0.0
psycopg2-binary==2.8.4
pyparsing==3.0.8
pytz==2019.3
redis==4.0.2
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
wrapt==1.14.0
zipp==3.8.0
